Checkpoints and multijump changes.

+++ MarballsUnity/Assets/Scripts/Level Scripts/KillZone.cs
@@ -2,7 +2,7 @@
 /// KillZone.cs
 /// Authors: Kyle Dawson, Charlie Sun, Chris Viqueira
 /// Date Created:  Feb. 16, 2015
-/// Last Revision: Apr. 22, 2015
+/// Last Revision: Apr. 27, 2015
 /// 
 /// Class that handles behavior of killzone boundaries.
 /// 
@@ -20,7 +20,7 @@ public class KillZone : MonoBehaviour {
 	
 	GameMaster gm;	// Reference to Game Master.
 
-	//public bool destroyMarble;	// Whether this killzone should obliterate the marble on contact or not.
+	public bool showGizmo = false;	// Whether gizmo should be rendered or not.
 
 	// Awake - Called before anything else. Use this to find the Game Master and tell it this exists.
 	void Awake () {
@@ -29,8 +29,7 @@ public class KillZone : MonoBehaviour {
 
 	// Use this for initialization
 	void Start () {
-		//if (GetComponent<Renderer>())
-		//	GetComponent<Renderer>().enabled = false;	// Kill zone should only be rendered in scene view, not in gameplay.
+
 	}
 	
 	// Update is called once per frame
@@ -50,8 +49,10 @@ public class KillZone : MonoBehaviour {
 
 	// OnDrawGizmosSelected - Used to draw things when selected in scene view exclusively: does not affect gameplay.
 	void OnDrawGizmosSelected() {
-		Gizmos.color = Color.red;
-		Gizmos.DrawCube(transform.position, transform.localScale);	// Shows kill zone when the kill zone is selected.
+		if (showGizmo) {
+			Gizmos.color = Color.red;
+			Gizmos.DrawCube(transform.position, transform.localScale);	// Shows kill zone when the kill zone is selected.
+		}
 	}
 
 	// Kill - Lives up to its namesake.
